[CI] Add a scheduled workflow for tests and benchmarks (#2822)
# Motivation

We want to run our unit tests, integration tests and benchmarks on a
regular schedule to understand how stable they are.

# Modification

This PR adds a scheduled workflow for the three above jobs. It runs it
three times a day so we get enough data throughout the day.

# Result

Continuous feedback about the stability of our tests.