Provide documentation and context information for NIOTooManyBytesError
The NIOTooManyBytesError doesn't have any documentation for someone
that encounters this error. Provide documentation that explains the
situation when the error occurs, which is the upTo limit of an
AsyncSequence is exceeded. Describe one potential action, which
is to increase this limit.

The error doesn't give the end user any sense of magnitude of the
byte limit that has been exceeded. Provide the maxBytes in the error
so that the user can gauge whether the upTo limit is already
reasonable and the payload size is excessive, or the limit needs
to be increased to suit more situations.