Align benchmark scaling and minimum samples
# Motivation

We want to have an aligned strategy across all of our benchmarks in how we use the scaling factor and samples.

# Modification

This PR makes sure that any inner loop is scaled by the scaling factor. Moreover, we want to run more than 1 iteration. To do this we set an almost infinite maximum duration and a maximum iteration count.

# Result

More consistent benchmarks