[CI] Run tests on push to main (#2868)
# Motivation

We want to get green check marks on our commits to main. To do this we
have to run part of our CI jobs when a new commit is pushed.

# Modification

This PR adds a new workflow trigger to run our CI when a commit is
pushed to the main branch.

# Result

Green checkmarks on our commits.