NIOSingleStepByteToMessageDecoder reentrancy safety
Motivation:

NIOSingleStepByteToMessageDecoder calls out part way through its
processing step to a user-provided closure which can cause re-entrant
behavior which violates the assumption made in the code that if the
buffer is non-empty at the start that will be true later in the method.

Modifications:

NIOSingleStepByteToMessageDecoder no longer assumes a non-empty buffer
in its final phase of buffer management.

Further changes to protect against re-entrancy shouldn't be necessary
because the outside call to `messageReceiver` is the only one which is
permitted to be re-entrant (`decode` and `decodeLast` are not).

Result:

NIOSingleStepByteToMessageDecoder can handle re-entrant processing
calls.