`FileChunkProducer` improved backpressure handling
In the original adoption of `NIOThrowingAsyncSequenceProducer` the code
did not deal with backpressure being applied very well.

A key bug was that it was not protected against re-entrant calls to `produceMore`. Such calls
may happen e.g. when the producer has been asked to pause producing and
then resume again before the initial read had completed. This resulted
in overlapping reads and re-ordered data.

This change introduces a new activity state which is protected by a lock
and keeps track of if we are in the critical section to serialize
access.

`DirectoryEntryProducer` performs most of its logic within a lock and so
doesn't seem to be impacted in the same way.