[CI] Add Windows matrix build
# Motivation

In the long term we want to add support for Windows in our Swift packages. To do this we need a CI setup to ensure that our packages are building and the tests are passing.

# Modification

This PR adds a windows job to our swift matrix that installs Swift on the runner and then uses Swift PM to build and test the package. Windows jobs are disabled by default since most of our packages require some work to add Windows support.

# Result

We have a Windows CI pipeline that allows us to verify that our packages are working.