Handle Sendability of RemovableChannelHandler
Motivation:

RemovableChannelHandlers have a large API surface in NIOCore.
That API surface is a bit awkward with regard to strict concurrency,
and needs some cleanup.

Modifications:

This patch adds some new API that is necessary to safely work with
RemovableChannelHandlers, deprecates some API that cannot plausibly
be used, and cleans up some other parts of the API.

Result:

Easier to work with RemovableChannelHandlers