BaseSocketChannel flushNow IONotificationState changes
Motivation:

We could previously hit an assert do to a re-entrancy issue where
channel and buffered write state could change during a call-out leading
to invalid state.

Modifications:

The decision on whether or not we should be registered for future writes
is now taken after the call outs to `fireChannelWritabilityChanged` and
`fireChannelReadComplete`. The new registration state is set to
`.unregister` if the channel is not open or if there are now flushed
pending writes.

Result:

Scope for re-entrancy crashes is reduced.

Co-authored-by: Cory Benfield <lukasa@apple.com>