Add isolated views to EventLoop, Promise, and Future
Motivation:

Users writing NIO code in a strict concurrency world often need to
interact with futures, promises, and event loops. The main interface
to these has strict sendability requirements, as it is possible the
user is doing so from outside the EventLoop that provides the isolation
domain for these types.

However, in many cases the user knows that they are on the isolation
domain in question. In that case, they need more capabilities. While
they can achieve their goals with NIOLoopBound today, it'd be nice if
they had a better option.

Modifications:

- Make EventLoop.Isolated public.
- Make EventLoopFuture.Isolated public.
- Make EventLoopPromise.Isolated public.
- Make all their relevant methods public.
- Move the runtime isolation check from the point of use to the point
    of construction.
- Make the types non-Sendable to ensure that isolation check is
    sufficient.
- Add unsafeUnchecked options to create these types when performance
    matters and correctness is clear.
- Add tests for their behaviour.
- Update the documentation.

Result:

Writing safe code with promises, futures, and event loops is
easier.