Better handle ECONNRESET on connected datagram sockets.
Motivation:

When a connected datagram socket sends a datagram to a port or
host that is not listening, an ICMP Destination Unreachable
message may be returned. That message triggers an ECONNRESET
to be produced at the socket layer.

On Darwin we handled this well, but on Linux it turned out that
this would push us into connection teardown and, eventually,
into a crash. Not so good.

To better handle this, we need to distinguish EPOLLERR from
EPOLLHUP on datagram sockets. In these cases, we should check
whether the socket error was fatal and, if it was not, we
should continue our execution having fired the error down the
pipeline.

Modifications:

Modify the selector code to distinguish reset and error.
Add support for our channels to handle errors.
Have most channels handle errors as resets.
Override the logic for datagram channels to duplicate the
logic in readable.
Add a unit test.

Result:

Better datagrams for all.