Update documentation comments
I've been reviewing some parts of `NIOFileSystem` and came across a
variety of formatting of documentation comments. I've adopted a stricter
100-character limit (arbitrary, really; it could be longer or shorter
and is totally not consistent with the rest of the repository). Also
updated a bunch of grammar in comments and made it easier to digest.
Fixed a few typos along the way.