Add Cxx interop swift settings to CI
Motivation:

At the moment the Cxx interoperability CI workflow doesn't actually test
Cxx interoperability.

Modifications:

The script now looks for the precise way that the `Package.swift`
currently formats the target entry and appends a `swiftSettings` entry
to it.

This approach is pretty brittle but has the advantage that it's the
same on Linux and Darwin. If this formatting changes too often then we
could give up on this and assume the availability of gnu-sed instead
(and make sure it's present on our CI containers).

Result:

Cxx interoperability checks will check what they are named for.