GHA: Separate matrix input handling (#3018)
### Motivation:

The new test matrix workflow is quite verbose in the most common case,
unit tests, outputting two workflow entries for handling and preparing
the matrix inputs.

In addition to this, it appeared that when running in `act` the matrix
preparation step would run multiple times.

### Modifications:

* Prioritize the current most common case, string inputs, and skip the
matrix preparation step in this case reducing the number of workflows
run.
* Separate the path for loading the JSON matrix from a file into its own
workflow.

### Result:

* Less noise in the output
* More reliable and efficient local matrix runs